requirements.txt
# requirements.txt
# Core dependencies
requests==2.32.3
beautifulsoup4==4.13.3
schedule==1.2.2
python-dotenv==1.0.1
python-telegram-bot==22.0
selenium==4.31.0
webdriver-manager==4.0.2

# ML and data processing
google-generativeai==0.8.4

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# Utilities
pytz>=2022.1